my-blog/web/requirements-dev.txt
-r requirements.txt
ipdb
pytest==6.1.2
pytest-django==4.1.0
pytest-mock==3.3.1
